+++ contracts/misc/UiPoolDataProviderV2.sol
@@ -60,7 +60,7 @@ contract UiPoolDataProviderV2 is IUiPoolDataProviderV2 {
 
     function getReservesList(
         ILendingPoolAddressesProvider provider,
-        uint8 trancheId
+        uint16 trancheId
     ) public view override returns (address[] memory) {
         ILendingPool lendingPool = ILendingPool(provider.getLendingPool());
         return lendingPool.getReservesList(trancheId);
@@ -68,7 +68,7 @@ contract UiPoolDataProviderV2 is IUiPoolDataProviderV2 {
 
     function getReservesData(
         ILendingPoolAddressesProvider provider,
-        uint8 trancheId
+        uint16 trancheId
     )
         public
         view
@@ -86,7 +86,7 @@ contract UiPoolDataProviderV2 is IUiPoolDataProviderV2 {
             reserveData.underlyingAsset = reserves[i];
 
             // reserve current state
-            // uint8 trancheId = uint8(i % DataTypes.NUM_TRANCHES);
+            // uint16 trancheId = uint8(i % DataTypes.NUM_TRANCHES);
             DataTypes.ReserveData memory baseData = lendingPool.getReserveData(
                 reserveData.underlyingAsset,
                 trancheId
@@ -195,7 +195,7 @@ contract UiPoolDataProviderV2 is IUiPoolDataProviderV2 {
 
     function getUserReservesData(
         ILendingPoolAddressesProvider provider,
-        uint8 trancheId,
+        uint16 trancheId,
         address user
     ) external view override returns (UserReserveData[] memory) {
         ILendingPool lendingPool = ILendingPool(provider.getLendingPool());
@@ -208,7 +208,7 @@ contract UiPoolDataProviderV2 is IUiPoolDataProviderV2 {
         );
 
         for (uint256 i = 0; i < reserves.length; i++) {
-            // uint8 trancheId = uint8(i % DataTypes.NUM_TRANCHES);
+            // uint16 trancheId = uint8(i % DataTypes.NUM_TRANCHES);
             DataTypes.ReserveData memory baseData = lendingPool.getReserveData(
                 reserves[i],
                 trancheId
