+++ contracts/misc/UiPoolDataProviderV2V3.sol
@@ -60,7 +60,7 @@ contract UiPoolDataProviderV2V3 is IUiPoolDataProviderV3 {
 
     function getReservesList(
         ILendingPoolAddressesProvider provider,
-        uint8 trancheId
+        uint16 trancheId
     ) public view override returns (address[] memory) {
         ILendingPool lendingPool = ILendingPool(provider.getLendingPool());
         return lendingPool.getReservesList(trancheId);
@@ -68,7 +68,7 @@ contract UiPoolDataProviderV2V3 is IUiPoolDataProviderV3 {
 
     function getReservesData(
         ILendingPoolAddressesProvider provider,
-        uint8 trancheId
+        uint16 trancheId
     )
         public
         view
@@ -90,7 +90,7 @@ contract UiPoolDataProviderV2V3 is IUiPoolDataProviderV3 {
             );
 
             // reserve current state
-            // uint8 trancheId = uint8(i % DataTypes.NUM_TRANCHES);
+            // uint16 trancheId = uint8(i % DataTypes.NUM_TRANCHES);
             DataTypes.ReserveData memory baseData = lendingPool.getReserveData(
                 reserveData.underlyingAsset,
                 trancheId
@@ -201,7 +201,7 @@ contract UiPoolDataProviderV2V3 is IUiPoolDataProviderV3 {
 
     function getUserReservesData(
         ILendingPoolAddressesProvider provider,
-        uint8 trancheId,
+        uint16 trancheId,
         address user
     ) external view override returns (UserReserveData[] memory, uint8) {
         ILendingPool lendingPool = ILendingPool(provider.getLendingPool());
@@ -214,7 +214,7 @@ contract UiPoolDataProviderV2V3 is IUiPoolDataProviderV3 {
         );
 
         for (uint256 i = 0; i < reserves.length; i++) {
-            // uint8 trancheId = uint8(i % DataTypes.NUM_TRANCHES);
+            // uint16 trancheId = uint8(i % DataTypes.NUM_TRANCHES);
             DataTypes.ReserveData memory baseData = lendingPool.getReserveData(
                 reserves[i],
                 trancheId
